Fixed tests and modified some functionality due to spring security

+++ src/main/java/io/sultanov/recipes/api/services/RecipeService.java
@@ -3,6 +3,7 @@ package io.sultanov.recipes.api.services;
 import io.sultanov.recipes.api.repos.RecipeRepository;
 import io.sultanov.recipes.exceptions.ObjectNotFoundException;
 import io.sultanov.recipes.models.Recipe;
+import io.sultanov.recipes.models.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -10,6 +11,7 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -17,8 +19,13 @@ import java.util.Optional;
 public class RecipeService {
     private final RecipeRepository recipeRepository;
 
+    private final AuthService authService;
+
+    private final UserService userService;
+
     public Long newRecipe(Recipe recipe) {
         recipe.setDate(LocalDateTime.now());
+        recipe.setAuthor(userService.findByEmail(authService.getAuthInfo().getUsername()));
         return recipeRepository.save(recipe).getId();
     }
 
@@ -27,12 +34,20 @@ public class RecipeService {
     }
 
     public void deleteRecipe(Long id) {
-        recipeRepository.findById(id).orElseThrow(ObjectNotFoundException::new);
+        Recipe recipe = recipeRepository.findById(id).orElseThrow(ObjectNotFoundException::new);
+        User user = userService.findByEmail(authService.getAuthInfo().getUsername());
+        if (!Objects.equals(user.getEmail(), recipe.getAuthor().getEmail())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied!");
+        }
         recipeRepository.deleteById(id);
     }
 
     public Recipe updateRecipe(Recipe recipeDto, Long id) {
         Recipe recipe = recipeRepository.findById(id).orElseThrow(ObjectNotFoundException::new);
+        User user = userService.findByEmail(authService.getAuthInfo().getUsername());
+        if (!Objects.equals(user.getEmail(), recipe.getAuthor().getEmail())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied!");
+        }
         recipe.setDate(LocalDateTime.now());
         recipe.setCategory(recipeDto.getCategory());
         recipe.setDirection(recipeDto.getDirection());
